Use better example images for versatile styles
Uses high res images and sets fixed width of images so they work on all
screens.

Also removes the overview table on the top of the page, because it is
difficult to keep this in sync with the rest. If we get more style we'll
have to rethink the layout of this page anway.

This is based on the PR #26 which has conflicts.